requirements.txt
# Core ML Dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# MLOps & Monitoring
mlflow>=2.5.0
prometheus-client>=0.17.0
psutil>=5.9.0

# Database & Configuration
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
pyyaml>=6.0